Sanele's Cooking Journey.
Sanele was invited to Lindiwe's birthday party. It turns out Lindiwe's parents have some excellent cooking skills and some secret family recipes. Chicken, beef, and leg of lamb were on the Braai, all cooked to perfection, not forgetting the roasted mielies. There was a big pot of pap with colorful chakalaka sauce next to it. There were salads, roast vegetables, and sweet potatoes as well as freshly baked bread straight from the oven. For dessert, there was malva pudding with custard and chocolate pudding with chocolate ice cream.  In this portfolio of evidence, you will develop a recipe app to start Sanele on his cooking journey.